+++ src/main/java/lucene4ir/RetrievalApp.java
@@ -1,7 +1,7 @@
 package lucene4ir;
 
-import lucene4ir.utils.TokenAnalyzerMaker;
-import org.apache.lucene.analysis.Analyzer;
+import lucene4ir.utils.RetrievalParams;
+import lucene4ir.utils.RetrievalParamsReader;
 import org.apache.lucene.document.Document;
 import org.apache.lucene.index.DirectoryReader;
 import org.apache.lucene.index.IndexReader;
@@ -11,17 +11,10 @@ import org.apache.lucene.search.IndexSearcher;
 import org.apache.lucene.search.Query;
 import org.apache.lucene.search.ScoreDoc;
 import org.apache.lucene.search.TopDocs;
-import org.apache.lucene.search.similarities.BM25Similarity;
-import org.apache.lucene.search.similarities.LMSimilarity.CollectionModel;
-import org.apache.lucene.search.similarities.Similarity;
+import org.apache.lucene.search.similarities.LMSimilarity;
 import org.apache.lucene.store.FSDirectory;
 
-import javax.xml.bind.JAXB;
-import javax.xml.bind.annotation.*;
 import java.io.*;
-import java.lang.reflect.InvocationTargetException;
-import java.util.ArrayList;
-import java.util.List;
 
 /**
  * Created by leif on 22/08/2016.
@@ -29,83 +22,11 @@ import java.util.List;
 public class RetrievalApp {
 
     private RetrievalParams params;
-    private Similarity similarityFunction;
+
     private IndexReader reader;
     private IndexSearcher searcher;
-    private Analyzer analyzer;
     private QueryParser parser;
-    private CollectionModel colModel;
-
-    private RetrievalParams readParamsFromFile(String paramFile) {
-        /*
-        Reads in the xml formatting parameter file
-        Maybe this code should go into the RetrievalParams class.
-
-        Actually, it would probably be neater to create a ParameterFile class
-        which these apps can inherit from - and customize accordinging.
-         */
-        RetrievalParams params = new RetrievalParams();
-
-        try {
-            params = JAXB.unmarshal(new File(paramFile), RetrievalParams.class);
-        } catch (Exception e) {
-            e.printStackTrace();
-            System.exit(1);
-        }
-
-        // TODO: Can't these just be default values (i.e. the constructor?)
-        if (params.maxResults == 0.0) {
-            params.maxResults = 1000;
-        }
-        if (params.b == 0.0) {
-            params.b = 0.75f;
-        }
-        if (params.beta == 0.0) {
-            params.beta = 500f;
-        }
-        if (params.k == 0.0) {
-            params.k = 1.2f;
-        }
-        if (params.delta == 0.0) {
-            params.delta = 1.0f;
-        }
-        if (params.lam == 0.0) {
-            params.lam = 0.5f;
-        }
-        if (params.mu == 0.0) {
-            params.mu = 500f;
-        }
-        if (params.c == 0.0) {
-            params.c = 10.0f;
-        }
-        if (params.model == null) {
-            params.model = new RetrievalParams.Model();
-            params.model.className = "org.apache.lucene.search.similarities.BM25Similarity";
-        }
-        if (params.runTag == null) {
-            params.runTag = params.model.className;
-        }
-
-        if (params.resultFile == null) {
-            params.resultFile = params.runTag + "_results.res";
-        }
-
-        System.out.println("Path to index: " + params.indexName);
-        System.out.println("Query File: " + params.queryFile);
-        System.out.println("Result File: " + params.resultFile);
-        System.out.println("Model: " + params.model.className);
-        System.out.println("Max Results: " + params.maxResults);
-        System.out.println("b: " + params.b);
-
-        if (params.tokenFilterFile != null) {
-            TokenAnalyzerMaker tam = new TokenAnalyzerMaker();
-            analyzer = tam.createAnalyzer(params.tokenFilterFile);
-        } else {
-            analyzer = LuceneConstants.ANALYZER;
-        }
-
-        return params;
-    }
+    private LMSimilarity.CollectionModel colModel;
 
     private void processQueryFile() {
         /*
@@ -150,6 +71,7 @@ public class RetrievalApp {
                 fw.close();
             }
         } catch (Exception e) {
+            e.printStackTrace();
             System.out.println(" caught a " + e.getClass() +
                     "\n with message: " + e.getMessage());
         }
@@ -177,47 +99,20 @@ public class RetrievalApp {
         return hits;
     }
 
-
-    private void setSimilarityFunction(RetrievalParams params) throws ClassNotFoundException, NoSuchMethodException, IllegalAccessException, InvocationTargetException, InstantiationException {
-        // use reflection to try to look up the class name
-        Class<?> func = Class.forName(params.model.className);
-
-        // create some objects to store the params to pass to the class
-        Object[] args = new Object[params.model.params.size()];
-        Class<?>[] types = new Class[params.model.params.size()];
-
-        // populate the collections with the params from the xml file
-        for (int i = 0; i < params.model.params.size(); i++) {
-            args[i] = params.model.params.get(i);
-            // TODO: The assumption is that the constructor uses the primitive float class
-            // (The `Float` class wraps the primitive `float` type)
-            types[i] = float.class;
-        }
-
-        // let's instantiate!
-        similarityFunction = (Similarity) func.getConstructor(types).newInstance(args);
-    }
-
     public RetrievalApp(String retrievalParamFile) throws IllegalAccessException {
         System.out.println("Retrieval App");
 
+        RetrievalParamsReader paramsReader = new RetrievalParamsReader();
         // attempt to build a model of the retrieval params from the xml file
-        params = readParamsFromFile(retrievalParamFile);
-
-        // use some of the params to set the similarity function
-        try {
-            setSimilarityFunction(params);
-        } catch (ClassNotFoundException | NoSuchMethodException | InvocationTargetException | InstantiationException e) {
-            e.printStackTrace();
-        }
+        params = paramsReader.read(retrievalParamFile);
 
         try {
             reader = DirectoryReader.open(FSDirectory.open(new File(params.indexName).toPath()));
             searcher = new IndexSearcher(reader);
 
-            searcher.setSimilarity(similarityFunction);
+            searcher.setSimilarity(paramsReader.getSimilarityFunction());
 
-            parser = new QueryParser("content", analyzer);
+            parser = new QueryParser("content", paramsReader.getAnalyzer());
 
         } catch (Exception e) {
             e.printStackTrace();
@@ -243,45 +138,6 @@ public class RetrievalApp {
         RetrievalApp retriever = new RetrievalApp(retrievalParamFile);
         retriever.processQueryFile();
 
-    }
-
-}
-
-
-@XmlRootElement
-@XmlAccessorType(XmlAccessType.FIELD)
-// TODO: Could this class be moved to it's own class (for reuse elsewhere?)
-class RetrievalParams {
-    public String indexName;
-    public String queryFile;
-    public String resultFile;
-    public int maxResults;
-    public float k;
-    public float b;
-    public float lam;
-    public float beta;
-    public float mu;
-    public float c;
-    public float delta;
-    public String runTag;
-    public String tokenFilterFile;
-
-    // the <model> element
-    @XmlElement(name = "model")
-    public Model model;
-
-    @XmlAccessorType(XmlAccessType.FIELD)
-    /**
-     * Encapsulate the className and parameters given to a similarity function
-     */
-    public static class Model {
-
-        @XmlAttribute
-        public String className;
-
-        @XmlAttribute(name = "params")
-        @XmlList
-        public List<Float> params = new ArrayList<>();
     }
 
 }
